[DDW-1156] Better naming
