+++ source/renderer/app/components/wallet/wallet-restore/StepDialogStyles.scss
