+++ source/renderer/app/components/wallet/wallet-restore/StepMnemonicsDialog.js
@@ -1,19 +1,19 @@
 // @flow
 import React, { Component } from 'react';
 import WalletRestoreDialog from './WalletRestoreDialog';
-import commonStyles from './WalletRestoreStyles.scss';
+import commonStyles from './StepDialogStyles.scss';
 
 type Props = {
   onContinue: Function,
   onClose: Function,
 };
 
-export default class WalletTypeDialog extends Component<Props> {
+export default class MnemonicsDialog extends Component<Props> {
   render() {
     const { onContinue, onClose } = this.props;
     return (
       <WalletRestoreDialog
-        stepNumber={0}
+        stepNumber={1}
         actions={[
           {
             primary: true,
@@ -23,7 +23,7 @@ export default class WalletTypeDialog extends Component<Props> {
         ]}
         onClose={onClose}
       >
-        <div className={commonStyles.component}>WALLET TYPE STEP CONTENT</div>
+        <div className={commonStyles.component}>MNEMONICS STEP CONTENT</div>
       </WalletRestoreDialog>
     );
   }
