+++ source/renderer/app/components/wallet/wallet-restore/StepWalletTypeDialog.js
@@ -1,19 +1,19 @@
 // @flow
 import React, { Component } from 'react';
 import WalletRestoreDialog from './WalletRestoreDialog';
-import commonStyles from './WalletRestoreStyles.scss';
+import commonStyles from './StepDialogStyles.scss';
 
 type Props = {
   onContinue: Function,
   onClose: Function,
 };
 
-export default class ConfigurationDialog extends Component<Props> {
+export default class WalletTypeDialog extends Component<Props> {
   render() {
     const { onContinue, onClose } = this.props;
     return (
       <WalletRestoreDialog
-        stepNumber={2}
+        stepNumber={0}
         actions={[
           {
             primary: true,
@@ -23,7 +23,7 @@ export default class ConfigurationDialog extends Component<Props> {
         ]}
         onClose={onClose}
       >
-        <div className={commonStyles.component}>CONFIGURATION STEP CONTENT</div>
+        <div className={commonStyles.component}>WALLET TYPE STEP CONTENT</div>
       </WalletRestoreDialog>
     );
   }
