+++ source/renderer/app/containers/wallet/dialogs/wallet-restore/StepMnemonicsContainer.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
-import SuccessDialog from '../../../../components/wallet/wallet-restore/SuccessDialog';
+import MnemonicsDialog from '../../../../components/wallet/wallet-restore/StepMnemonicsDialog';
 import type { InjectedDialogContainerStepProps } from '../../../../types/injectedPropsType';
 import { InjectedDialogContainerStepDefaultProps } from '../../../../types/injectedPropsType';
 
@@ -10,11 +10,11 @@ const DefaultProps = InjectedDialogContainerStepDefaultProps;
 
 @inject('stores', 'actions')
 @observer
-export default class SuccessDialogContainer extends Component<Props> {
+export default class MnemonicsDialogContainer extends Component<Props> {
   static defaultProps = DefaultProps;
 
   render() {
     const { onClose, onContinue } = this.props;
-    return <SuccessDialog onClose={onClose} onContinue={onContinue} />;
+    return <MnemonicsDialog onClose={onClose} onContinue={onContinue} />;
   }
 }
